experiments/requirements.txt
google-genai>=1.56
